% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Generates a list of fatalities}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector (or matrix) used to represent the years of the data files to be imported.}
}
\value{
A ist with each element containing the Month and Year for each dataset imported. A warning  message,
  "invalid year: xxxx" will be displayed if the file does not exist.
}
\description{
The function generates a list summarising the number of fatal traffic accidents recorded in the US National Highway Traffic
  Safety Administration's Fatality Analysis Reporting System (FARS). Each file imported (representing data for a specific
  year) will be represented as a tibble with column headings of Month and Year.
}
\details{
The function iterates over each year using lapply to import the files created using the make_filename() function
  supplied in this package. The mutute and select functions from the dplyr package are also used to generate the
  required output.
}
\note{
A package permitting the use of pipelines ( %>% ) such as magrittr must be installed and attached.
  The dplyr package must be installed in the R environment but does not have to be attached.
}
\examples{
fars_read_years (2014)
fars_read_years(c(2013,2015))

}
