% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary by year and month of the number of fatal accidents recorded in the US NHTSA FARS system}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector (or matrix) used to represent the years of the data files to be imported.}
}
\value{
This function returns a tibble summarising the numbers of fatal accidents by month and year.
}
\description{
The function generates a tibble summarising the number of fatal injuries suffered in motor vehicle accidents recorded by the
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System (FARS). The tibble shows the number of
accidents grouped by month and year.
}
\details{
The function accepts a \code{years} argument and converts the elements from the list generated from the
fars_read_years() function into a tibble which summarises the number of fatalities incurred in motor vehicle
accidents by month and year.
}
\note{
A package permitting the use of pipelines ( %>% ), such as magrittr must be installed and attached.
  The dplyr and tidyr packages must be installed in the R environment but do not have to be attached.
}
\examples{
fars_summarize_years (2013)
fars_summarize_years (c(2013,2014))

}
