% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A value used to represent the year. This can be in either character or numeric format as the function will
try to coerce the value to an integer.}
}
\value{
A filename representing the fatal traffic incident data for a given year. The file will be of the form;
 accident_xxxx.csv.bz2 where xxxx is the year of data being imported.
}
\description{
Generate the name of a file containing traffic fatality statistics provided by the US National Highway Traffic Safety
  Administration's Fatality Analysis Reporting System (FARS).
}
\details{
The function accepts a year argument and uses the sprintf function to return a filename containing accident data
for the \code{year} parameter.
}
\examples{
make_filename(2013)

}
